+++ src/lib/config.ts
@@ -23,6 +23,8 @@ export const spotifyConfig = {
   scopes: [
     "playlist-modify-private",
     "playlist-modify-public",
+    "playlist-read-collaborative",
+    "playlist-read-private",
     "user-library-modify",
     "user-read-private"
   ]